Fix tests and shift functions around
