+++ tests/test_dockertools.py
@@ -151,70 +151,3 @@ def test_check_docker_size_toobig():
          pytest.raises(ValueError,
                        match="Docker image must be less than a terabyte."):
         dockertools.check_docker_size(response)
-
-
-def test_validate_docker_submission_valid():
-    """Tests that True is returned when validate docker
-    submission is valid"""
-    config = Mock()
-    docker_repo = 'docker.synapse.org/syn1234/docker_repo'
-    docker_digest = 'sha123566'
-    username = str(uuid.uuid1())
-    password = str(uuid.uuid1())
-    docker_sub = synapseclient.Submission(evaluationId=1,
-                                          entityId=1,
-                                          versionNumber=2,
-                                          dockerRepositoryName=docker_repo,
-                                          dockerDigest=docker_digest)
-    with patch.object(SYN, "getConfigFile", return_value=config),\
-         patch.object(config, "items",
-                      return_value={'username': username,
-                                    'password': password}),\
-         patch.object(SYN, "getSubmission",
-                      return_value=docker_sub) as patch_get_sub,\
-         patch.object(dockertools, "validate_docker",
-                      return_value=True) as patch_validate:
-        valid = dockertools.validate_docker_submission(SYN, "123455")
-        patch_validate.assert_called_once_with(docker_repo="syn1234/docker_repo",
-                                               docker_digest=docker_digest,
-                                               index_endpoint=ENDPOINT_MAPPING['synapse'],
-                                               username=username,
-                                               password=password)
-        patch_get_sub.assert_called_once_with("123455")
-        assert valid
-
-
-def test_validate_docker_submission_nousername():
-    """Tests ValueError is thrown when no username or password is passed"""
-    config = Mock()
-    password = str(uuid.uuid1())
-    with patch.object(SYN, "getConfigFile", return_value=config),\
-         patch.object(config, "items",
-                      return_value={'username': None,
-                                    'password': password}),\
-         pytest.raises(ValueError,
-                       match='Synapse config file must have username and password'):
-        dockertools.validate_docker_submission(SYN, "123455")
-
-
-def test_validate_docker_submission_notdocker():
-    """Tests ValueError is thrown when submission is not docker submission"""
-    config = Mock()
-    docker_repo = 'docker.synapse.org/syn1234/docker_repo'
-    docker_digest = None
-    username = str(uuid.uuid1())
-    password = str(uuid.uuid1())
-    docker_sub = synapseclient.Submission(evaluationId=1,
-                                          entityId=1,
-                                          versionNumber=2,
-                                          dockerRepositoryName=docker_repo,
-                                          dockerDigest=docker_digest)
-    with patch.object(SYN, "getConfigFile", return_value=config),\
-         patch.object(config, "items",
-                      return_value={'username': username,
-                                    'password': password}),\
-         patch.object(SYN, "getSubmission",
-                      return_value=docker_sub),\
-         pytest.raises(ValueError,
-                       match='Submission is not a Docker submission'):
-        dockertools.validate_docker_submission(SYN, "123455")
